LPS-186331 split view changes toolbar from results to so we can refactor results to frontend dataset display
