+++ modules/apps/change-tracking/change-tracking-web/src/main/resources/META-INF/resources/init.jsp
@@ -69,6 +69,7 @@ page import="com.liferay.portal.kernel.util.WebKeys" %>
 <%@ page import="java.text.Format" %>
 
 <%@ page import="java.util.Date" %><%@
+page import="java.util.Map" %><%@
 page import="java.util.Objects" %>
 
 <%@ page import="javax.portlet.RenderResponse" %>
